+++ example/example.yaml
@@ -1,0 +1,1 @@
+#sigma
